+++ crates/stackable-shared/src/lib.rs
@@ -2,4 +2,7 @@
 //! workspace.
 
 pub mod crd;
+pub mod secret;
+
+pub mod time;
 pub mod yaml;